Add files via upload
OK I like that you listed the file name and the line of code! Everything else it there, I just need you to sort it slightly differently. See my marked up copy for an example...I basically just want the screen to be the last string of the namespace e.g. In ScreenDriver::DialogChangeBag, I really only want the "DialogChangeBag" and then all possible variations of that screen. You could even do them in an array if you didn't want to repeat the screen, but I'm lazy so I would just do it this way.